BLT.Binding

BLT Solver API

bltInit

Arguments

:: RatMode
mode for handing high precision rationals

-> Bool
flag to enable Yices "sound mode" in BLT

-> IO HContext

Create a new problem context

Each call to bltInit must be paired with a call to bltFree when its problem context is no longer needed.

bltVar :: HContext -> String -> IO CExpr

Create a fresh variable with given name in the context

bltConst :: HContext -> Rational -> IO CExpr

Create a constant term

bltAdd :: HContext -> CExpr -> CExpr -> IO CExpr

Add two expressions

bltSmul :: HContext -> Rational -> CExpr -> IO CExpr

Multiply an expression by a scalar. The scalar is checked for overflow before being passed to the C-API.

bltAssume

Arguments

:: HContext
context pointer

-> Rational
left hand side (lower bound)

-> CExpr
linear middle term

-> Rational
right hand side (upper bound)

-> IO BLTCode
return code

Push a new assumption onto the stack

bltCheck :: HContext -> IO BLTCode

Check the current assumption stack for satisfiability

bltModel :: HContext -> String -> IO (Maybe Int64)

Get the value of a variable in the solver's model

Return the value assigned to variable with given name or Nothing if there is no value or if the variable doesn't exist in the given context.

bltSave :: HContext -> IO CSolvState

Save the current solver state

Return an opaque pointer to the current assumption stack so that it may be backtracked to later using bltBacktrack.

bltBacktrack :: HContext -> CSolvState -> IO BLTCode

Load a previously saved solver state

Backtrack the current context to the state represented by the given CSolvState value. This operation removes assumptions from the stack but preserves the list of known variables.

bltFree :: HContext -> IO ()

Free memory associated with the given context

data HContext

A problem context

data RatMode

Choice of modes for handling rational coefficients

Constructors

RatID
pass rat'l coefficients through to BLT

RatApprox Int
use an approximation strategy which limits the precision of denominators

RatFixed Int64
use a fixed precision strategy which scales all coefficients uniformly and then rounds to the nearest integer (towards zero on ties)

re-exports from BLT.Internal

type CExpr = Ptr Expr

type CSolvState = Ptr SolvState

data BLTCode

Type wrapper for BLT return codes defined in "codes.h"

BLT return codes

bltSAT :: BLTCode

Rename some constants of type BLTCode for better compatibility

bltUNSAT :: BLTCode